Update save to eeprom functionality After testing and calibration, the date, serial number and tuning parameters are saved in the eeprom. Functions dealing with eeprom were moved in a separate file.
